Converted support to float32 and fixed requirements

+++ model/pytorch/dcrnn_cell.py
@@ -78,7 +78,7 @@ class DCGRUCell(torch.nn.Module):
         # this is to ensure row-major ordering to equal torch.sparse.sparse_reorder(L)
         indices = indices[np.lexsort((indices[:, 0], indices[:, 1]))]
         L = torch.sparse_coo_tensor(indices.T, L.data, L.shape, device=device)
-        return L
+        return L.float()
 
     def forward(self, inputs, hx):
         """Gated recurrent unit (GRU) with Graph Convolution.
